stdenv bootstrap: upgrade x86_64-linux bootstrap tools & use temporary new files
We cannot use the previous bootstrap files for the stdenv stage0
bootstrap as these are using glibc-2.27 and then `gcc` fails during a
link phase with the following error:

    version `GLIBC_2.33' not found (required by /build/build/./gcc/liblto_plugin.so)

A workaround for that is to disable LTO (`enableLTO = false` for the
gcc10 drv), however we don't want that in stage3 (where this is needed
for the fix to work) as we'd then have a `stdenv.cc` without link-time
optimization enabled.

To work around that I decided to rebuild the bootstrap tools and pushed
the products on my employer's fileshare - I guess Mayflower is
sufficiently trustworthy to build that on hydra.nixos.org. Of course,
this is only a temporary measure until everything else is done and we
can push them to `tarballs.nixos.org`.

I added a (relatively hacky) `config.inBootstrap = true;` nixpkgs-option
that indicates that `nixpkgs` is instantiated in order to build
bootstrap files. With this option set to `true`, `enableLTO` will be set
to `false`.

Also, this is only a POC, updates for `aarch64-linux` and `i686-linux`
will follow.